feat: Move metrics rules and alerts to separate package
Moved the definitions of alerts and rules
to pkg/monitoring/rules package.
