+++ tools/prom-metrics-collector/metrics_collector.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	parser "github.com/kubevirt/monitoring/pkg/metrics/parser"
-	"kubevirt.io/ssp-operator/internal/operands/metrics"
-
 	dto "github.com/prometheus/client_model/go"
+
+	"kubevirt.io/ssp-operator/pkg/monitoring/rules"
 )
 
 // This should be used only for very rare cases where the naming conventions that are explained in the best practices:
@@ -14,7 +14,7 @@ var excludedMetrics = map[string]struct{}{}
 
 func readMetrics() []*dto.MetricFamily {
 	var metricFamilies []*dto.MetricFamily
-	sspMetrics := metrics.RecordRulesDescList
+	sspMetrics := rules.RecordRulesDescList
 
 	for _, metric := range sspMetrics {
 		if _, isExcludedMetric := excludedMetrics[metric.Name]; !isExcludedMetric {
